(frontend) Changing DeviceSettings and AddDevice into a single component named « DevicesManagement ».
This commit is done in order to reduce code redundancy with add and modify device components into a single component named « DevicesManagement ».
