--- web/src/components/DeviceSettings.tsx
@@ -1,123 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import FormDeviceOptions from './FormDeviceOptions.tsx'
-import {Modal, Button, Form, Image, Alert} from 'react-bootstrap'
-import { RiSettings2Line } from "react-icons/ri";
-
-class DeviceSettings extends React.Component {
-    constructor(props){
-        super(props)
-        this.state = {
-            modalShow: false,
-            setModalShow: false,
-            validAlertShow: false,
-            invalidAlertShow: false,
-            setInvalidAlertShow: false,
-            setValidAlertShow: false,
-            messageErreur: "",
-        }
-    };
-
-    handleClose() {
-        this.setState({modalShow: false});
-        this.setState({validAlertShow: false});
-    }
-    handleShow() {
-        this.setState({modalShow: true});
-    }
-    handleValidAlertShow() {
-        if (this.invalidAlertShow) {
-            this.setState({validAlertShow: false});
-        } else {
-            this.setState({validAlertShow: true});
-            this.setState({invalidAlertShow: false});
-        }
-    }
-
-    handleInvalidAlertShow() {
-        if (this.validAlertShow) {
-            this.setState({invalidAlertShow: false})
-        } else {
-            this.setState({validAlertShow: false})
-            this.setState({invalidAlertShow: true})
-        }
-    }
-    
-    handleSettingsInput = (e, id) => {
-        e.preventDefault();
-
-        let dataDevice = {
-            "groupId": 1,
-            "name" : e.target.elements.formNameDevice.value.trim(),
-            "modele": 1,
-            "type": 1,
-            "ip": e.target.elements.formIpAddress.value.trim(),
-        }
-        let regexIpv4Address = /^(?:[0-9]{1,3}\.){3}[0-9]{1,3}$/;
-        let valueIpv4AddressForm = dataDevice.ip;
-
-        if(((e.target.elements.formNameDevice.value).trim() == "") && (valueIpv4AddressForm).length > 15 || !(regexIpv4Address.test(valueIpv4AddressForm))) {
-            this.handleInvalidAlertShow();
-            this.state.messageErreur = "le nom de l'appareil doit contenir au moins un caractère ou le format de l'adresse IP n'est pas valide.";
-        }
-        else if ((e.target.elements.formNameDevice.value).trim() == "") {
-            this.handleInvalidAlertShow();
-            this.state.messageErreur = "le nom de l'appareil doit contenir au moins un caractère.";
-         } 
-        else if ((valueIpv4AddressForm).length > 15 || !(regexIpv4Address.test(valueIpv4AddressForm))) {
-            this.handleInvalidAlertShow();
-            this.state.messageErreur = "le format de l'adresse IP n'est pas valide.";
-         }
-
-        else {
-
-         fetch("https://localhost/api/devices/" + id, {method: 'PUT',
-            headers: {
-                'Content-type': 'application/json; charset=UTF-8'
-            },
-             body: JSON.stringify(dataDevice)
-         })
-             .then(response => response.json())
-             .then(this.props.displayDevice.bind(this))
-    }
-}
-    
-    render (){
-        return (
-            <div>
-                <Button variant="secondary" onClick={this.handleShow.bind(this)} >
-                    <RiSettings2Line/>
-                </Button>
-            
-                <Modal show={this.state.modalShow} onHide={this.handleClose.bind(this)} animation={true}>
-                    <Form onSubmit={(e) => this.handleSettingsInput(e, this.props.id)}>
-                        <Modal.Header closeButton>
-                            <Modal.Title>Paramètres de l'appareil</Modal.Title>
-                        </Modal.Header>
-
-                        <Modal.Body> 
-                            <FormDeviceOptions name={this.props.name} ipAddress={this.props.ipAddress} id={this.props.id} groupId={this.props.groupId} type={this.props.type} displayDevice={this.props.displayDevice} />
-                            <Alert variant="success" show={this.state.validAlertShow}>
-                                La configuration de l'appareil « {this.props.name} » a  bien été sauvegardée.
-                            </Alert>
-                            <Alert variant="danger" show={this.state.invalidAlertShow}>
-                                La configuration de l'appareil « {this.props.name} » ne peut pas être sauvegardée car {this.state.messageErreur}
-                            </Alert>
-                        </Modal.Body>
-                    
-                        <Modal.Footer>
-                            <Button variant="secondary" onClick={this.handleClose.bind(this)}>
-                                Fermer la fenêtre
-                            </Button>
-                            <Button variant="primary" onClick={this.handleValidAlertShow.bind(this)} type="submit">
-                                Sauvegarder
-                            </Button>
-                        </Modal.Footer>
-                    </Form>
-                </Modal>
-            </div>
-        );
-  };
-}
-
-export default DeviceSettings;
